Abstract out the indexing expression as a substitution rule. (#28)
This patch rewrites the transformation machinery in `feinsum` such that the arguments can come
in only as substitution rule invocations. To change this, we had to change code-generation, and,
matching pathways in feinsum. The substitutions enable passing in expressions lowered from a 
lazy-evaluated multi-dimensional IR as arguments of the einsum. The matching logic also supports
impure array accesses coming in that are inferred simply from the expression dependencies of the
multiplicative terms in the sum-reduction.

To simplify the matching logic we needed a data-structure for bidirectional dictionaries, and hence we
also introduce a dependency on bidict.

This patch, however, is lean on the documentation describing the grammar of the expressions that
`feinsum` will handle and telling the user about the matching logic in a transparent sense. We do not
handle that over here to keep this PR manageable for review.